chore(quality): break bots/utils barrel import cycle, enforce circular lane
madge flagged one cycle: libs/shared/ts/src/bots/utils/index.ts <-> streaming.ts.
index.ts was both a barrel (re-exporting streaming) and the home of the text
helpers (chunkResponse, truncateResponse, PLATFORM_LIMITS, ...). streaming.ts
imported those helpers back from the barrel, closing the loop.

Extract the text helpers into a dedicated text.ts. index.ts becomes a pure
barrel (export * from "./text"); streaming.ts imports straight from ./text.
Public surface of @gaia/shared is unchanged.

Flip the `circular` lane to enforced in quality-gate.
